of_net: generalize the use of_get_mac_mode()
The `of_get_mac_mode()` is similar to the `of_get_phy_mode()` function, but
it reads the 'mac-mode' device-tree property (vs 'phy-mode' &
'phy-connection-type').

This is typically useful when there is a passive mode converter in-between
the MAC & PHY (e.g. GMII-to-RGMII converter) and the MAC needs to be
configured in one mode, and PHY in another.

The 'stmmac' driver had implemented a similar function (called
`stmmac_of_get_mac_mode()`). Now, the intent is to implement a similar
function for the Cadence macb driver.

This change moves the stmmac driver's function into `of_net`, to be re-used
from there.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>