net: macb: implement support for mac-mode
In some setups, there can be a passive mode converter between MAC & PHY
(e.g. MII-to-RMII). In this case the MAC needs to be configured for MII and
PHY for RMII.

This change adds support to the MACB driver to read the 'mac-mode' property
and use that the configure the MAC's interface to the PHY. If unspecified,
the old PHY interface is used (as before).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>