iio: buffer,event: duplicate chardev creation for buffers & events
This patch moves the chardev creation into industrialio-buffer.c &
industrialio-event.c. The move has to be in a single step (which makes the
patch a bit big), in order to pass the reference of the chardev to
'indio_dev->chrdev'.

For structure purposes, industrialio-core.c should be the place where
cdev_device_add()/device_add() gets called, and the deletion function as
well.

What happens after this patch is:
- 'indio_dev->chrdev' is converted to a pointer
- if there is an IIO buffer, it will pass a reference to it's chardev
  object to 'indio_dev->chrdev'
- if it doesn't, the event interface will attach a reference to it's
  chardev (if it is instantiated)

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>