iio: buffer: rework sysfs creation for buffer & scan_elements directories
This change reworks the creation of the 2 directories required for IIO
buffers ('buffer' & 'scan_elements'). It's an intermediate rework of the
'iio_buffer_alloc_sysfs_and_mask()' function.

This rework requires that 'iio_buffer_alloc_sysfs_and_mask()' be called
after 'cdev_device_add()', because it's required that the IIO device be
registered, i.e. the kobject (of the IIO device) be initialized, in order
to attach new kobjects (files, groups & symlinks).

This also has a side-effect, that we can reduce IIO_MAX_GROUPS by 2, which
affects the size of 'struct iio_dev'.

With this change, adding more sysfs/attribute groups to an IIO device
becomes more flexible, since it is not limited by 'IIO_MAX_GROUPS' anymore.

We also need to add explicit sysfs_ops for the 2 directories now, to unpack
the objects in a way that the current hooks still work.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>