iio: core: fail early in iio_device_alloc() if we can't get a device id
This change moves the 'ida_simple_get()' call to be the first one in
iio_device_alloc(). It cleans up the error path a bit as we don't need to
call any kfree(dev) anymore. We allocate an IIO device only if we have
managed to obtain a device ID.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>