iio: core: use new common ioctl() mechanism
This change makes use of the new centralized ioctl() mechanism. The event
interface registers it's ioctl() handler to IIO device, and calls the
iio_device_ioctl() function to execute it.

Similarly the chardev for the IIO buffer calls the iio_device_ioctl(), so
if the event interface has registered an ioctl() it can get executed.

This mechanism will allow for other ioctl() commands to be registered, from
other parts of the IIO framework, and they should be callable via the
iio_device_ioctl() call.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>