iio: core: use new common ioctl() mechanism
This change makes use of the new centralized ioctl() mechanism. The event
interface registers it's ioctl() handler to IIO device.
Both the buffer & event interface call 'iio_device_ioctl()', which should
take care of all of indio_dev's ioctl() calls.

Later, we may add per-buffer ioctl() calls, and since each buffer will get
it's own chardev, the buffer ioctl() handler will need a bit of tweaking
for the first/legacy buffer (i.e. indio_dev->buffer).
Also, those per-buffer ioctl() calls will not be registered with this
mechanism.

The move of the comment in this patch is intentional to show that [with
this patch], the 'indio_dev->info' NULL check is still around, but it just
got moved.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>