spi: spi-axi-spi-engine: Update write size
With the previous changes in 1f542798e5b, a variable write size feature
was added.
In this context, the spi engine can write a word with a size between 8 and
32. This commit changes the buff pointer inside the
spi_engine_offload_load_msg from uint8_t to uint32_t in order to write
more than 8 bits at a time.

Signed-off-by: Sergiu Cuciurean <sergiu.cuciurean@analog.com>