jesd204: add topology ID and link ID support in DT
This should cover cases of a device belonging in multiple JESD204
topologies and support multiple JESD204 links per topology.

Now the 'jesd204-top-device' is an u32 which is the ID of the topology.
A top-level device must also provide a list of link IDs via
'jesd204-link-ids'.

Then each 'jesd204-inputs' list must contain 2 identifiers: the ID of the
topology that must be matched with 'jesd204-top-device' and a JESD204 link
ID that must match one of the IDs in the 'jesd204-link-ids'.

So, something like:
  jesd204-top-device = <0>;
  jesd204-link-ids = <0 1>;
  jesd204-inputs = <input 0 0>, <input 0 1>;

The JESD204 link IDs will be assigned to 'struct jesd204_link', so that if
needed they can be accessed from that object.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>